Load more button
Make load more button. It's a last major think that I wanna try on this page. But I have some more ideas how to show my skills on this web page. I think they will be done in couple of days cause i already have mark up for this stuff it must to be done!